+++ zstacklib/zstacklib/utils/linux.py
@@ -94,7 +94,10 @@ def rm_dir_force(dpath, only_check=False):
     if dpath.strip() in black_dpath_list:
         raise Exception("how dare you delete directory %s" % dpath)
     if os.path.exists(dpath) and not only_check:
-        shutil.rmtree(dpath)
+        if os.path.isdir(dpath):
+            shutil.rmtree(dpath)
+        else:
+            rm_file_force(dpath)
     else:
         return dpath
